Add custom hook for user accounts
